+++ src/main/java/chess/controller/MovePositionDto.java
@@ -1,7 +1,7 @@
 package chess.controller;
 
-import chess.domain.board.Position;
 import chess.domain.piece.Color;
+import chess.domain.position.Position;
 
 public record MovePositionDto(Position sourcePosition, Position targetPosition, Color color) {
