+++ src/main/java/chess/domain/board/InitialBoardGenerator.java
@@ -1,14 +1,15 @@
 package chess.domain.board;
 
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.NoPiece;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.slidingpiece.Bishop;
+import chess.domain.piece.slidingpiece.Queen;
+import chess.domain.piece.slidingpiece.Rook;
+import chess.domain.position.Position;
 import java.util.HashMap;
 import java.util.Map;
