+++ src/main/java/chess/domain/piece/slidingpiece/SlidingPiece.java
@@ -1,7 +1,8 @@
-package chess.domain.piece;
+package chess.domain.piece.slidingpiece;
 
-import chess.domain.board.Direction;
-import chess.domain.board.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 import java.util.Map;
 import java.util.Set;
