+++ src/main/java/chess/view/OutputView.java
@@ -1,15 +1,15 @@
 package chess.view;
 
 import chess.domain.board.Board;
-import chess.domain.board.Position;
-import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.NoPiece;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.slidingpiece.Bishop;
+import chess.domain.piece.slidingpiece.Queen;
+import chess.domain.piece.slidingpiece.Rook;
+import chess.domain.position.Position;
 
 public class OutputView {
