+++ src/test/java/chess/domain/piece/EmptyBoardGenerator.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import static chess.domain.piece.Color.NO_COLOR;
 
-import chess.domain.board.Position;
+import chess.domain.position.Position;
 import java.util.HashMap;
 import java.util.Map;
